requirements.txt
fastapi
uvicorn
faiss-cpu
sqlalchemy
sqlite3
pydantic
python-dotenv